Replace 'x' image with unicode text and fix font scaling
Following a fix of the font scaling the font used to signal that
a file had not been loaded was too large.
The /img/x.png was also not part of the library and relied on the
client app to have it. It's not easy to build an image into a library
so opted to use a unicode character insteaad which yields the same
effect.